Fix return type for OSX clipboard
I was using a model where an outer `Result` (representing
issues accessing the clipboard) contained an inner `Option`
(representing whether a value is in the clipboard). This
was not the model used by the rest of the code, where the
`Result` covers both cases.

I think that model makes more sense, so I've converted to
that. OTOH, I've changed some Errs in osx_clipboard.rs to
panic!s, because they indicated programming errors.